Refactor new stories into separate fixtures/components
